Move ConsumeData files to consumeData folder
