+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/consumeDataEventStreams/ResourcesSection.tsx
