+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/consumeDataEventStreams/index.tsx
