+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/consumeDataRestAPI/Credentials/credentialsDataGrid/generateCredentialsDialog/GeneratedCredentialsSection.tsx
