+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/consumeDataRestAPI/DisableRestApiDialog.tsx
@@ -10,10 +10,10 @@ import { DialogContentText } from '@mui/material';
 
 import { Dialog } from '@dolittle/design-system';
 
-import { useConnectionsIdRestApiDisablePost } from '../../../../apis/integrations/connectionRestApiApi.hooks';
-import { CACHE_KEYS } from '../../../../apis/integrations/CacheKeys';
+import { useConnectionsIdRestApiDisablePost } from '../../../../../apis/integrations/connectionRestApiApi.hooks';
+import { CACHE_KEYS } from '../../../../../apis/integrations/CacheKeys';
 
-import { useConnectionIdFromRoute } from '../../../routes.hooks';
+import { useConnectionIdFromRoute } from '../../../../routes.hooks';
 
 export type DisableRestApiDialogProps = {
     isOpen: boolean;
